refactor(eql-scalars): split lib.rs into definitions + impl/test modules
The crate's single lib.rs had grown to 1215 lines, ~55% of it tests, with each
enum's definition interleaved with its impl block. Reorganise for readability so
the catalog reads top-to-bottom:

- lib.rs (~240 lines) keeps the *definitions* — the six type defs, the two
  crate-internal macros (fixtures!, int_values!), and all catalog data
  (ORDERED_INT_DOMAINS, INT*_FIXTURES, INT4/INT2/INT8/DATE, CATALOG, INT*_VALUES).
- Inherent impls move to sibling modules: kind.rs (BoundedIntKind, ScalarKind),
  term.rs (Term), fixture.rs (Fixture), spec.rs (ScalarSpec). Methods travel with
  their types, so no re-exports are needed.
- The 7 unit-test modules move verbatim into tests.rs (#[cfg(test)] mod tests),
  kept as one group because rust_tests spans multiple types; use super::* becomes
  use crate::*.

Pure code-move, zero behaviour change. Public API surface of the crate is
unchanged. Verified: cargo test -p eql-scalars (40 pass), codegen:parity
(byte-for-byte identical int4 golden), test:crates clippy -D warnings clean,
eql-codegen/eql-tests-macros build unchanged.
