+++ crates/eql-scalars/src/kind.rs
@@ -1,0 +1,100 @@
+//! Inherent impls for the scalar-kind vocabulary: [`BoundedIntKind`] (the total
+//! accessors for fixed-width integer kinds) and [`ScalarKind`] (the native
+//! scalar a domain maps onto). Definitions live in `lib.rs`.
+
+use crate::{BoundedIntKind, ScalarKind};
+
+impl BoundedIntKind {
+    /// The Rust type name as it appears in generated source (e.g. `"i32"`).
+    pub const fn rust_type(self) -> &'static str {
+        match self {
+            BoundedIntKind::I16 => "i16",
+            BoundedIntKind::I32 => "i32",
+            BoundedIntKind::I64 => "i64",
+        }
+    }
+
+    /// The `MIN` named-constant symbol (e.g. `"i32::MIN"`).
+    pub const fn min_symbol(self) -> &'static str {
+        match self {
+            BoundedIntKind::I16 => "i16::MIN",
+            BoundedIntKind::I32 => "i32::MIN",
+            BoundedIntKind::I64 => "i64::MIN",
+        }
+    }
+
+    /// The `MAX` named-constant symbol (e.g. `"i32::MAX"`).
+    pub const fn max_symbol(self) -> &'static str {
+        match self {
+            BoundedIntKind::I16 => "i16::MAX",
+            BoundedIntKind::I32 => "i32::MAX",
+            BoundedIntKind::I64 => "i64::MAX",
+        }
+    }
+
+    /// The zero literal symbol (always `"0"`).
+    pub const fn zero_symbol(self) -> &'static str {
+        "0"
+    }
+
+    /// Inclusive lower bound of the representable range, widened to `i128`.
+    pub const fn min_value(self) -> i128 {
+        match self {
+            BoundedIntKind::I16 => i16::MIN as i128,
+            BoundedIntKind::I32 => i32::MIN as i128,
+            BoundedIntKind::I64 => i64::MIN as i128,
+        }
+    }
+
+    /// Inclusive upper bound of the representable range, widened to `i128`.
+    pub const fn max_value(self) -> i128 {
+        match self {
+            BoundedIntKind::I16 => i16::MAX as i128,
+            BoundedIntKind::I32 => i32::MAX as i128,
+            BoundedIntKind::I64 => i64::MAX as i128,
+        }
+    }
+}
+
+impl ScalarKind {
+    /// The fixed-width integer kinds — those with `i128` bounds and
+    /// `Min`/`Max`/`Zero` sentinels — projected onto [`BoundedIntKind`], or
+    /// `None` for the non-integer kinds. The single boundary where "this kind has
+    /// bounds" is decided; the bounded accessors live on `BoundedIntKind` and are
+    /// total there. NOT an orderability test: `Numeric`/`Text`/`Date` are
+    /// ORE-orderable yet not integers.
+    pub const fn as_bounded_int(self) -> Option<BoundedIntKind> {
+        match self {
+            ScalarKind::I16 => Some(BoundedIntKind::I16),
+            ScalarKind::I32 => Some(BoundedIntKind::I32),
+            ScalarKind::I64 => Some(BoundedIntKind::I64),
+            ScalarKind::Numeric | ScalarKind::Text | ScalarKind::Jsonb | ScalarKind::Date => None,
+        }
+    }
+
+    /// True for the fixed-width integer kinds. Gates the bounded-numeric
+    /// invariants. Equivalent to `self.as_bounded_int().is_some()`.
+    pub const fn is_int(self) -> bool {
+        self.as_bounded_int().is_some()
+    }
+
+    /// True for chrono-backed temporal kinds (`Date`; `Timestamptz` once added) —
+    /// the kinds whose test `ScalarType` impl is generated by `temporal_values!`
+    /// rather than the integer proc-macro path. Replaces the `[temporal]` marker.
+    pub const fn is_temporal(self) -> bool {
+        matches!(self, ScalarKind::Date)
+    }
+
+    /// The Rust type name as it appears in generated source (e.g. `"i32"`).
+    pub const fn rust_type(self) -> &'static str {
+        match self {
+            ScalarKind::I16 => "i16",
+            ScalarKind::I32 => "i32",
+            ScalarKind::I64 => "i64",
+            ScalarKind::Numeric => "numeric",
+            ScalarKind::Text => "text",
+            ScalarKind::Jsonb => "jsonb",
+            ScalarKind::Date => "chrono::NaiveDate",
+        }
+    }
+}
